I.6 : kBefore/kAfter en uint256 dans swapWrapper (correctif pre-existant)
reserves() rend uint72 ; uint72 * uint72 panique (Solidity 0.8,
overflow) des que les reserves depassent ~6,8e10 sats. swapWrapper
revert alors avant assertGe(kAfter, kBefore) : le controle local de
non-decroissance de k ne tournait jamais au-dela de ce seuil, et une
assertion qui panique avant de s'executer donne une fausse confiance.

Cast explicite uint256() aux deux lignes, rien d'autre. La
non-decroissance de k reste portee au global par
invariant_shareValueNeverDecreases (deja en uint256 sur la somme des
reserves). Bug pre-existant du fichier, releve a la relecture Opus
III.2.

test/Pool.invariant.t.sol : 7 passing.

+++ backend/test/Pool.invariant.t.sol
@@ -146,7 +146,12 @@ contract PoolHandler is CommonBase, StdUtils, StdAssertions {
     uint256 amount = bound(_amount, 1, 21_000_000e8);
     uint256 minOut = bound(_minOut, 0, expectedSwapAmountOut(indexIn, amount, indexOut));
 
-    uint256 kBefore = pool.reserves(indexIn) * pool.reserves(indexOut);
+    // uint256() explicite : reserves() rend uint72, et uint72 * uint72
+    // panique (Solidity 0.8, overflow) des que les reserves depassent
+    // ~6,8e10 sats. Sans le cast, swapWrapper revert avant
+    // assertGe(kAfter, kBefore) et le controle local de non-decroissance
+    // de k ne tourne jamais sur les gros etats.
+    uint256 kBefore = uint256(pool.reserves(indexIn)) * uint256(pool.reserves(indexOut));
     address mgrAtSwap = pool.manager();
     uint256[3] memory owedBefore = feeRegistrySnapshot(mgrAtSwap);
     (uint256[3] memory reservesBefore, uint256[3] memory balancesBefore) = poolState();
@@ -196,7 +201,7 @@ contract PoolHandler is CommonBase, StdUtils, StdAssertions {
       swapsUnderManager++;
     }
 
-    uint256 kAfter = pool.reserves(indexIn) * pool.reserves(indexOut);
+    uint256 kAfter = uint256(pool.reserves(indexIn)) * uint256(pool.reserves(indexOut));
     assertGe(kAfter, kBefore);
   }
